+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Changes
 - renaming the threads spawned for deploy / destroy to indicate the module being worked on
 - adding detailed docs for CLI parameters
+- complete support to delete the seedkit on `seedfarmer destroy` command
 
 ### Fixes